% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsigma_2.R
\name{correct_pol}
\alias{correct_pol}
\title{correct_pol}
\usage{
correct_pol(data, px = 0, py = 0, angle = 0)
}
\arguments{
\item{data}{Input \code{tibble} in \code{sigma_2}/\code{fsigma_2}-compatible format.}

\item{px}{Correction to \code{Px}.}

\item{py}{Correction to \code{Py}.}

\item{angle}{Correction to angle.}
}
\value{
Updated table.
}
\description{
correct_pol
}
