% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_names.R
\name{fix_names}
\alias{fix_names}
\title{fix_names}
\usage{
fix_names(text)
}
\arguments{
\item{text}{Names of the CSV file to replace}
}
\value{
R-compatible names suitable for further work
}
\description{
fix_names
}
