% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_Q.R
\name{generate_Q}
\alias{generate_Q}
\title{generate_Q}
\usage{
generate_Q(p_x, p_y, w_x, w_y, p_x_mean, p_y_mean)
}
\arguments{
\item{p_x}{Vector of stokes parameters}

\item{p_y}{Vector of stoke parameters}

\item{w_x}{Vector of weights}

\item{w_y}{Vector of weights}

\item{p_x_mean}{Weighted average stokes parameter}

\item{p_y_mean}{Weighted average stokes parameter}
}
\value{
Covariance matrix
}
\description{
Generates covariance matrix from Dipol data
}
