% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_test.R
\name{h_test}
\alias{h_test}
\title{h_test}
\usage{
h_test(data, p_x = Px, p_y = Py, sg = SG, cov = Q, n = N)
}
\arguments{
\item{data}{Two-row input tibble}

\item{p_x}{Px polarization column name}

\item{p_y}{Py polarization column name}

\item{sg}{Sigma column name}

\item{cov}{Covariance column name}

\item{n}{Number of observations column name}
}
\description{
Performs Hotelling T^2 test on the Dipol-2 averages
}
