% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_test.R
\name{h_test2}
\alias{h_test2}
\title{h_test2}
\usage{
h_test2(left, right, ..., id, px = "Px", py = "Py", n = "N", q = "Q")
}
\arguments{
\item{left, right}{Equal-sized \code{data.frame}s for pairwise testing}

\item{...}{Additional columns to include to the output.
Value from the \code{left} \code{data.frame} is included.}

\item{id}{Id column to match entries. If absent, row numbers are used
(\code{tibble::rowid_to_column}).}

\item{px, py}{Column that contains value of Stokes \code{q} and \code{u},
present in both tables/}

\item{n}{Column that contains number of data points used to average.}

\item{q}{Column that contains the specific covariance matrix
(as \code{list} of \code{matrix})}
}
\value{
A \code{data.frame} with test results
}
\description{
h_test2
}
