% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_from_legacy_descriptor.R
\name{load_from_legacy_descriptor}
\alias{load_from_legacy_descriptor}
\title{load_from_legacy_descriptor}
\usage{
load_from_legacy_descriptor(desc, root = ".")
}
\arguments{
\item{desc}{Descriptors as output from \code{read_legacy_descriptor}.}

\item{root}{Root for the files in the descriptor.}
}
\description{
load_from_legacy_descriptor
}
