% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadFromLegacyDescriptor.R
\name{LoadFromLegacyDescriptor}
\alias{LoadFromLegacyDescriptor}
\title{LoadFromLegacyDescriptor}
\usage{
LoadFromLegacyDescriptor(desc, root = ".")
}
\arguments{
\item{desc}{Descriptors as output from \code{ReadLegacyDescriptor}.}

\item{root}{Root for the files in the descriptor.}
}
\description{
LoadFromLegacyDescriptor
}
