% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_path.R
\name{process_files}
\alias{process_files}
\title{process_files}
\usage{
process_files(path, by = 4L)
}
\arguments{
\item{path}{Path to the files. Supports globbing}

\item{by}{Average by how many}
}
\value{
A list of tibbles
}
\description{
process_files
}
