% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_legacy_descriptor.R
\name{read_legacy_descriptor}
\alias{read_legacy_descriptor}
\title{Read legacy descriptor}
\usage{
read_legacy_descriptor(path)
}
\arguments{
\item{path}{Path to the descriptor file (e.g. \code{Vin.txt})}
}
\description{
Read legacy descriptor
}
