% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadLegacyDescriptor.R
\name{ReadLegacyDescriptor}
\alias{ReadLegacyDescriptor}
\title{ReadLegacyDescriptor}
\usage{
ReadLegacyDescriptor(path)
}
\arguments{
\item{path}{Path to the descriptor file (e.g. \code{Vin.txt})}
}
\description{
ReadLegacyDescriptor
}
