% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_2.R
\name{Sigma_2}
\alias{Sigma_2}
\alias{sigma_2}
\title{Sigma_2}
\usage{
Sigma_2(data, filter = "B", bandInfo = NULL, ...,
  eqtrialCorrFactor = 0.034907, ittMax = 500L, eps = 1e-16,
  date = JD, obs = Obs)
}
\arguments{
\item{data}{Input data in form of a \code{tibble}.}

\item{filter}{Filter name.}

\item{bandInfo}{A \code{tibble} containing
filter descriptions.}

\item{...}{Extra variables that should be preserved. Method preserves first element of the column}

\item{eqtrialCorrFactor}{Equatorial correction for the angle.}

\item{ittMax}{Maximum iterations to perform.}

\item{eps}{Tolerance. If average difference between previous and current step
is smaller than \code{eps}, execution stops.}

\item{date}{Name of the date column. Supports \code{rlang}
quasiquotiong.}

\item{obs}{Name of the observations column. Supports \code{rlang}
quasiquotiong.}
}
\description{
Sigma_2
}
