% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigma_2.R
\name{sigma_2}
\alias{sigma_2}
\alias{Sigma_2}
\title{sigma_2}
\usage{
sigma_2(
  data,
  filter = "B",
  band_info = NULL,
  ...,
  itt_max = 500L,
  eps = 1e-16,
  date = JD,
  obs = Obs,
  eqtrialCorrFactor = lifecycle::deprecated(),
  bandInfo = lifecycle::deprecated(),
  ittMax = lifecycle::deprecated()
)

Sigma_2(
  data,
  filter = "B",
  band_info = NULL,
  ...,
  itt_max = 500L,
  eps = 1e-16,
  date = JD,
  obs = Obs,
  eqtrialCorrFactor = lifecycle::deprecated(),
  bandInfo = lifecycle::deprecated(),
  ittMax = lifecycle::deprecated()
)
}
\arguments{
\item{data}{Input data in form of a \code{tibble}.}

\item{filter}{Filter name.}

\item{band_info}{A \code{tibble} containing
filter descriptions.}

\item{...}{Extra variables that should be preserved. Method preserves first element of the column}

\item{itt_max}{Maximum iterations to perform.}

\item{eps}{Tolerance. If average difference between previous and current step
is smaller than \code{eps}, execution stops.}

\item{date}{Name of the date column. Supports \code{rlang}
quasiquotiong.}

\item{obs}{Name of the observations column. Supports \code{rlang}
quasiquotiong.}

\item{eqtrialCorrFactor}{\emph{Deprecated since 0.4.1 and no longer used}.}

\item{bandInfo}{\emph{Deprecated since 0.4.1 in favor of \code{band_info}}.}

\item{ittMax}{\emph{Deprecated since 0.4.1 in favor of \code{itt_max}}.}
}
\description{
sigma_2
}
